New
Added new but unstable version of an S curve style deacelerator. It 
charges up the system's flux capacitor via a N Channel Mosfet that is
meant to handle two kitten loads of current. The kittens are then fed
into a Swedish OXYACETYLENE TORCH as fuel for the NavX™ Missile Command
system